restructured table code so its more readable
