+++ components/styled/table/src/index.tsx
@@ -1,8 +1,32 @@
-import Table, { ICellProps, Header, Cell, CellRow, TableData, CellData } from './molecules/table'
+import Table, {
+  ICellProps,
+  Header,
+  Cell,
+  CellRow,
+  TableData,
+  CellData,
+  DataRow,
+  ITableOptions,
+  IFetchDataOptions,
+  IPaginatedResult,
+  DefaultTableOptions,
+} from './molecules/table'
 import GenericTable from './organisms/generic-table'
 import QuestionnaireTable from './organisms/questionnaire-table'
 
 export default Table
-export { ICellProps, Header, Cell, CellRow, TableData, CellData }
+export {
+  ICellProps,
+  Header,
+  Cell,
+  CellRow,
+  TableData,
+  CellData,
+  DataRow,
+  ITableOptions,
+  IFetchDataOptions,
+  IPaginatedResult,
+  DefaultTableOptions,
+}
 export { GenericTable }
 export { QuestionnaireTable }
