+++ components/styled/table/src/organisms/questionnaire-table.tsx
@@ -1,10 +1,10 @@
 import { QuestionnaireResponse, Axis, Questionnaire } from '@ltht-react/types'
 import { FC, useMemo } from 'react'
 import { Icon } from '@ltht-react/icon'
-import Table from '../molecules/table'
+import Table, { IFetchDataOptions, IPaginatedResult, ITableOptions } from '../molecules/table'
 import mapQuestionnaireDefinitionAndResponsesToTableData from './questionnaire-table-methods'
 
-const QuestionnaireTable: FC<IProps> = ({ definition, records, headerAxis = 'y' }) => {
+const QuestionnaireTable: FC<IProps> = ({ definition, records, headerAxis = 'y', ...props }) => {
   const tableData = useMemo(() => mapQuestionnaireDefinitionAndResponsesToTableData(definition, records, headerAxis), [
     headerAxis,
     definition,
@@ -22,13 +22,15 @@ const QuestionnaireTable: FC<IProps> = ({ definition, records, headerAxis = 'y' 
     )
   }
 
-  return <Table tableData={tableData} />
+  return <Table tableData={tableData} {...props} />
 }
 
 interface IProps {
   definition: Questionnaire
   records: QuestionnaireResponse[]
   headerAxis?: Axis
+  tableOptions?: ITableOptions
+  fetchData?: (options: IFetchDataOptions) => Promise<IPaginatedResult>
 }
 
 export default QuestionnaireTable
